Ajout de l'embed anime "complet" dans un autre fichier
